requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15.0
django-extensions>=3.2.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
Pillow>=11.0.0
